feat(eon-com): breaking change to oclif
