--- .eslintrc.js
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2020, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-
-module.exports = {
-  extends: ["eslint-config-salesforce-typescript"],
-  parser: "@typescript-eslint/parser",
-};
