+++ .prettierrc.json
@@ -1,1 +1,9 @@
-"@salesforce/prettier-config"
+{
+  "printWidth": 120,
+  "trailingComma": "es5",
+  "useTabs": false,
+  "tabWidth": 4,
+  "semi": true,
+  "singleQuote": true,
+  "bracketSpacing": true
+}
