--- src/commands/eon/package/devkit/create.ts
@@ -1,123 +1,0 @@
-/*
- * Copyright (c) 2020, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-import * as os from 'os';
-import { flags, SfdxCommand } from '@salesforce/command';
-import fs from 'fs';
-import path from 'path';
-import { Messages, NamedPackageDir, SfdxProjectJson } from '@salesforce/core';
-
-import { AnyJson } from '@salesforce/ts-types';
-import EONLogger, { COLOR_ERROR, COLOR_HEADER, COLOR_INFO } from '../../../../eon/EONLogger';
-import { LOGOBANNER } from '../../../../eon/logo';
-import {
-  DEVKITFOLDER,
-  EONDEVKITYML,
-  EXAMPLEAPEX,
-  EXPORTJSON,
-  SCRIPTSFOLDER,
-  SETUPFILE,
-  TESTDATAFOLDER,
-} from '../../../../helper/devkit-constants';
-// Initialize Messages with the current plugin directory
-Messages.importMessagesDirectory(__dirname);
-
-// Load the specific messages for this file. Messages from @salesforce/command, @salesforce/core,
-// or any library that is using the messages framework can also be loaded this way.
-const messages = Messages.loadMessages('@eon-com/eon-sfdx', 'devkit');
-
-export default class Create extends SfdxCommand {
-  public static description = messages.getMessage('commandDescription');
-
-  public static examples = messages.getMessage('examples').split(os.EOL);
-
-  public static args = [{ name: 'file' }];
-
-  protected static flagsConfig = {
-    package: flags.string({
-      char: 'p',
-      description: messages.getMessage('packageFlag'),
-      required: true,
-    }),
-  };
-
-  // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
-  protected static requiresProject = true;
-  public async run(): Promise<AnyJson> {
-    EONLogger.log(COLOR_HEADER(LOGOBANNER));
-    const packagename = this.flags.package;
-
-    const projectJson: SfdxProjectJson = await this.project.retrieveSfProjectJson();
-
-    let packageDirs: NamedPackageDir[] = projectJson.getUniquePackageDirectories();
-
-    // check if package exists
-    if (!packageDirs.map((a) => a.package).includes(packagename)) {
-      EONLogger.log(COLOR_ERROR('Package ' + packagename + ' does not exist in sfdx.project.json'));
-      return;
-    }
-    const packagePath: string = packageDirs.find((a) => a.package === packagename).path;
-    const filePaths: String[] = this.findFileInDir(packagePath, SETUPFILE);
-
-    if (filePaths.length > 1) {
-      EONLogger.log(COLOR_ERROR('Multiple ' + SETUPFILE + ' files found in package ' + packagename));
-      EONLogger.log(COLOR_ERROR('Only one ' + SETUPFILE + ' file is allowed per package'));
-      return;
-    } else if (filePaths.length === 1) {
-      EONLogger.log(COLOR_INFO(SETUPFILE + ' file in package ' + packagename + ' already exists.'));
-      return;
-    }
-
-    // create new subfolders of devkit
-    if (!fs.existsSync(path.join(packagePath, DEVKITFOLDER))) {
-      fs.mkdirSync(path.join(packagePath, DEVKITFOLDER));
-    }
-    if (!fs.existsSync(path.join(packagePath, DEVKITFOLDER, SCRIPTSFOLDER))) {
-      fs.mkdirSync(path.join(packagePath, DEVKITFOLDER, SCRIPTSFOLDER));
-    }
-    if (!fs.existsSync(path.join(packagePath, DEVKITFOLDER, TESTDATAFOLDER))) {
-      fs.mkdirSync(path.join(packagePath, DEVKITFOLDER, TESTDATAFOLDER));
-    }
-    // create new yml file
-    const baseStructureYaml = EONDEVKITYML;
-    if (!fs.existsSync(path.join(packagePath, DEVKITFOLDER, SETUPFILE))) {
-      fs.writeFileSync(path.join(packagePath, DEVKITFOLDER, SETUPFILE), baseStructureYaml);
-      EONLogger.log(COLOR_INFO('Created ' + SETUPFILE + ' file in package ' + packagename));
-    }
-    // create new export.json file
-    const baseExportJson = EXPORTJSON;
-
-    if (!fs.existsSync(path.join(packagePath, DEVKITFOLDER, TESTDATAFOLDER, 'export.json'))) {
-      fs.writeFileSync(path.join(packagePath, DEVKITFOLDER, TESTDATAFOLDER, 'export.json'), baseExportJson);
-      EONLogger.log(COLOR_INFO('Created export.json file in package ' + packagename));
-    }
-    // create example apex script
-    const exampleApexBody = EXAMPLEAPEX;
-    if (!fs.existsSync(path.join(packagePath, DEVKITFOLDER, SCRIPTSFOLDER, 'setup-script.apex'))) {
-      fs.writeFileSync(path.join(packagePath, DEVKITFOLDER, SCRIPTSFOLDER, 'setup-script.apex'), exampleApexBody);
-      EONLogger.log(COLOR_INFO('Created setup-script.apex file in package ' + packagename));
-    }
-
-    EONLogger.log(COLOR_HEADER('Devkit creation completed for package ' + packagename));
-    return {};
-  }
-
-  // find all files in directory
-  findFileInDir(dir, filename) {
-    let results = [];
-    fs.readdirSync(dir).forEach((file) => {
-      let fullPath = path.join(dir, file);
-      if (fs.lstatSync(fullPath).isDirectory()) {
-        results = [...results, ...this.findFileInDir(fullPath, filename)];
-      } else {
-        if (fullPath.includes(filename)) {
-          results.push(fullPath);
-        }
-      }
-    });
-    return results;
-  }
-}
