+++ src/utils/get-packages.ts
@@ -1,8 +1,8 @@
-import { SfdxProjectJson } from '@salesforce/core';
+import { SfProjectJson } from '@salesforce/core';
 import { PackageTree } from '../interfaces/package-interfaces';
 import { NamedPackageDirLarge } from '../helper/types';
 
-export function getDeployUrls(projectJson: SfdxProjectJson, packagename: string): PackageTree {
+export function getDeployUrls(projectJson: SfProjectJson, packagename: string): PackageTree {
   const packageDirs: NamedPackageDirLarge[] = projectJson.getUniquePackageDirectories();
   let packageTree: PackageTree;
   const currentPackage: NamedPackageDirLarge = packageDirs.find((pck) => pck.package === packagename);
