--- test/mocha.json
@@ -1,7 +1,0 @@
-{
-  "require": "ts-node/register,source-map-support/register",
-  "watch-extensions": "ts",
-  "recursive": true,
-  "reporter": "spec",
-  "timeout": 5000
-}
