--- test/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@salesforce/dev-config/tsconfig-test",
-  "include": ["./**/*.ts"],
-  "compilerOptions": {
-    "skipLibCheck": true
-  }
-}
